requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.4.4
typing_extensions==4.11.0
boto3==1.34.29
botocore==1.34.29
dj-database-url==1.0.0
django-cors-headers==3.12.0
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
drf-nested-routers==0.93.5
jmespath==1.0.1
pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
django-quill-editor==0.1.33
mysqlclient==2.2.4